(new) Re-implementing a API endpoint file, in Paws.c.
This file acts as the primary bond between API consumers and the `Paws.c` codebase. It includes practically everything else, defines quite a few core structures, and is basically essential in all circumstances.

This began as a duplication of the old Paws.c, but I changed it so heavily that I decided this commit didn’t deserve the (dup) label.