(new api) Implemented the `fork` nuketype.
This also includes the `ll` and `element` structs, and families for all of the above.

This code was initially copied from old-world Paws, and subsequently modified massively. I don’t think I’m going to try to do this all in one commit again. I’ll probably directly duplicate old work next time, with an (incomplete) label, and then work towards making it up-to-date, then close the changes with a (complete:) label.

This also still includes some old, irrelevant code. The code has been updated to match current style standards, and to compile (and pass the old tests)… and that’s about it. More work to be done here.