(- re style) Moving the `initialize` functions into the “family functions” section.
Semantically speaking, `initialize()` operates on `struct a_type *`, not `a_type` itself. Until it’s been initialized, it’s not really an object of that type, now is it? Since it doesn’t operate on fully-operational instances of `a_type`, I’m considering them a family function instead of an instance function.